removed dead code, moved some functions to mod.rs in python dir, removed py_sighash.rs as it wasn't needed, added access to a sha256d function
