+++ python/src/tx_engine/__init__.py
@@ -3,9 +3,8 @@ This is a tx_engine doc str
 """
 # noqa: F401 - 'x' - imported but unused
 
-from tx_engine.tx_engine import Tx, TxIn, TxOut, Script, Wallet, p2pkh_script, hash160, address_to_public_key_hash, public_key_to_address  # noqa: F401
-from tx_engine.tx_engine import SigHashCache # noqa: F401
-from tx_engine.tx_engine import partial_sig_hash, full_sig_hash, sig_hash_preimage #noqa: F401
+from tx_engine.tx_engine import Tx, TxIn, TxOut, Script, Wallet, p2pkh_script, hash160, hash256d, address_to_public_key_hash, public_key_to_address  # noqa: F401
+from tx_engine.tx_engine import sig_hash_preimage, sig_hash #noqa: F401
 from tx_engine.engine.context import Context  # noqa: F401
 from tx_engine.engine.util import encode_num, decode_num  # noqa: F401
 from tx_engine.tx.sighash import SIGHASH #noqa: F401
