+++ src/python/hashes.rs
@@ -14,3 +14,9 @@ pub fn hash160(data: &[u8]) -> Vec<u8> {
     let sha256 = Sha256::digest(data);
     Ripemd160::digest(sha256).to_vec()
 }
+
+/// Hashes a data array twice using SHA256
+pub fn sha256d(data: &[u8]) -> Vec<u8> {
+    let sha256 = Sha256::digest(data);
+    Sha256::digest(sha256).to_vec()
+}
